Median Updates - Maintain AVL Balance
Despite calling this an AVL, the data structure has come far in being something
more unique. It's not based on height, and has drifted away from the traditional
AVL implementation of methods such as add. This change introduces new logic
to add where it ensures that the new node being added is put into the smaller
of either side of the root node. Pivoting is not as essential as retrieving the
node smallest on the right or largest on the left to replace the root before adding
a new member to the tree.